cruft removal, renaming files, restructuring contests
