+++ src/components/ContestTile.js
